move bscAddress to bsc folder
