+++ test/bsc/xWinFixture.js
@@ -12,7 +12,8 @@ const {
   deployxWinMasterChef,
   deployLockStaking,
 } = require("./xWinTestHelpers.js");
-const { bsc, hardhatNode } = require("../../bscMainnetAddresses.js");
+const { bsc, hardhatNode } = require("./bscMainnetAddresses.js
+");
 const { ethers } = require("hardhat");
 
 async function xWinFixture() {
