+++ test/bsc/xWinPriceMaster.js
@@ -6,7 +6,7 @@ const { anyValue } = require("@nomicfoundation/hardhat-chai-matchers/withArgs");
 const { expect } = require("chai");
 const { xWinFixture } = require("./xWinFixture");
 const { expectAlmostEquals } = require("./xWinTestHelpers.js");
-const { bsc, hardhatNode } = require("../../bscMainnetAddresses.js");
+const { bsc, hardhatNode } = require("./bscMainnetAddresses.js");
 const { ethers } = require("hardhat");
 
 describe("xWinPriceMaster", function () {
